requirements.txt
Flask
gunicorn
numpy
sklearn
pandas
PyYAML
six
evidently
dill
-e .